remove spotless rename based package
