+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'Winter'
+rootProject.name = 'dependency-injection-helper'
