+++ src/main/java/com/patryklikus/dihelper/beans/BeanInitializationException.java
@@ -1,0 +1,16 @@
+/* Copyright patryklikus.com All Rights Reserved. */
+package com.patryklikus.dihelper.beans;
+
+public class BeanInitializationException extends RuntimeException {
+    public BeanInitializationException(Throwable cause) {
+        super(cause);
+    }
+
+    public BeanInitializationException(String message) {
+        super(message);
+    }
+
+    public BeanInitializationException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
