+++ src/main/java/com/patryklikus/dihelper/lifecycle/Close.java
@@ -1,0 +1,16 @@
+/* Copyright patryklikus.com All Rights Reserved. */
+package com.patryklikus.dihelper.lifecycle;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+import static java.lang.annotation.ElementType.METHOD;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(METHOD)
+public @interface Close {
+    boolean enabled() default true;
+
+    short order() default 0;
+}
