+++ src/main/java/com/patryklikus/dihelper/lifecycle/Initable.java
@@ -1,0 +1,9 @@
+/* Copyright patryklikus.com All Rights Reserved. */
+package com.patryklikus.dihelper.lifecycle;
+
+public interface Initable {
+    /**
+     * Prepares the object to make it usable.
+     */
+    void init();
+}
