+++ src/main/java/com/patryklikus/dihelper/searcher/ClassSearchException.java
@@ -1,0 +1,12 @@
+/* Copyright patryklikus.com All Rights Reserved. */
+package com.patryklikus.dihelper.searcher;
+
+public class ClassSearchException extends RuntimeException {
+    ClassSearchException(String message) {
+        super(message);
+    }
+
+    ClassSearchException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
