--- src/main/java/com/patryklikus/winter/lifecycle/Init.java
@@ -1,16 +1,0 @@
-/* Copyright patryklikus.com All Rights Reserved. */
-package com.patryklikus.winter.lifecycle;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-import static java.lang.annotation.ElementType.METHOD;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(METHOD)
-public @interface Init {
-    boolean enabled() default true;
-
-    short order() default 0;
-}
