--- src/main/java/com/patryklikus/winter/lifecycle/Initable.java
@@ -1,9 +1,0 @@
-/* Copyright patryklikus.com All Rights Reserved. */
-package com.patryklikus.winter.lifecycle;
-
-public interface Initable {
-    /**
-     * Prepares the object to make it usable.
-     */
-    void init();
-}
