--- src/main/java/com/patryklikus/winter/searcher/ClassSearchException.java
@@ -1,12 +1,0 @@
-/* Copyright patryklikus.com All Rights Reserved. */
-package com.patryklikus.winter.searcher;
-
-public class ClassSearchException extends RuntimeException {
-    ClassSearchException(String message) {
-        super(message);
-    }
-
-    ClassSearchException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
