+++ src/test/java/com/patryklikus/dihelper/beans/exampleProject/Main.java
@@ -1,0 +1,7 @@
+/* Copyright patryklikus.com All Rights Reserved. */
+package com.patryklikus.dihelper.beans.exampleProject;
+
+public class Main {
+    public static void main(String[] args) {
+    }
+}
