+++ src/test/java/com/patryklikus/dihelper/beans/exampleProject/models/Apple.java
@@ -1,0 +1,5 @@
+/* Copyright patryklikus.com All Rights Reserved. */
+package com.patryklikus.dihelper.beans.exampleProject.models;
+
+public class Apple implements Fruit {
+}
