--- src/test/java/com/patryklikus/winter/beans/exampleProject/Main.java
@@ -1,7 +1,0 @@
-/* Copyright patryklikus.com All Rights Reserved. */
-package com.patryklikus.winter.beans.exampleProject;
-
-public class Main {
-    public static void main(String[] args) {
-    }
-}
