--- src/test/java/com/patryklikus/winter/beans/exampleProject/models/Fruit.java
@@ -1,5 +1,0 @@
-/* Copyright patryklikus.com All Rights Reserved. */
-package com.patryklikus.winter.beans.exampleProject.models;
-
-public interface Fruit {
-}
